create single bridge provider context
